Add support for CLIENT SETNAME command to cpp_redis::Subscriber (#22)
* Allow 64bit ints for TTL in client.
  Redis server supports 64 bit ints for TTL.
  Actually up to 0x7FFFFFFFFFFFFFFFLL milliseconds.
  Or 0x7FFFFFFFFFFFFFFFLL/1000 seconds.

* If we lose the TCP conenction and it is then reconnected, there was a problem were the AUTH
command got serialized twice to the buffer to be sent to the TCP connection by unprotected_auth:
  Once when it was directly added to the buffer AND
  Once because it was added to the command queue it got serialize to the buffer again
However there was only one call back on the command queue.  So we got 2 responces and only
one callback to handle them.  So the next command in the queue gets the second AUTH response
instead of the expected responce and everything is off by one.

* Added support for CLIENT SETNAME in cpp_redis::subscriber